+++ apps/frontend/src/app/2_molecules/NativeTokenAmount/NativeTokenAmount.tsx
@@ -1,6 +1,9 @@
 import React, { FC, useMemo } from 'react';
 
+import { ChainId } from '@sovryn/ethers-provider';
+
 import { getCurrencyPrecision } from '../../5_pages/PortfolioPage/components/ProtocolSection/ProtocolSection.utils';
+import { useCurrentChain } from '../../../hooks/useChainStore';
 import { useGetNativeTokenPrice } from '../../../hooks/useGetNativeTokenPrice';
 import { decimalic } from '../../../utils/math';
 import { AmountRenderer } from '../AmountRenderer/AmountRenderer';
@@ -9,14 +12,20 @@ type NativeTokenAmountProps = {
   usdValue?: string | number;
   dataAttribute?: string;
   precision?: number;
+  chainId?: ChainId;
 };
 
 export const NativeTokenAmount: FC<NativeTokenAmountProps> = ({
   usdValue,
   dataAttribute,
   precision,
+  chainId,
 }) => {
-  const { price: nativeTokenPrice, nativeToken } = useGetNativeTokenPrice();
+  const currentChainId = useCurrentChain();
+
+  const { price: nativeTokenPrice, nativeToken } = useGetNativeTokenPrice(
+    chainId ?? currentChainId,
+  );
 
   const value = useMemo(
     () =>