+++ apps/frontend/src/app/3_organisms/Header/Header.tsx
@@ -74,7 +74,7 @@ export const Header: FC = () => {
                 text={t(translations.header.nav.spicePoints)}
                 style={ButtonStyle.primary}
                 className="bg-[#24BFB74D]/[0.3] border-[#24BFB74D]/[0.3] hover:bg-[#24BFB74D]"
-                onClick={() => navigate('/claim-lp')}
+                onClick={() => navigate('/bob-lp-points')}
               />
             )}
           </ol>